Rename `numba.cuda.runtime` to `memory_management`
The name `numba.cuda.runtime` as a package aliased the already-existing
`Runtime` object. Although the package name is now no longer consistent
with the package from core Numba containing the Numba Runtime code, it
needs to be renamed so that we can avoid clashing with already-extant
CUDA names.

Fixes #174
