+++ numba_cuda/numba/cuda/memory_management/__init__.py
@@ -1,0 +1,1 @@
+from numba.cuda.memory_management.nrt import rtsys  # noqa: F401
