--- numba_cuda/numba/cuda/runtime/__init__.py
@@ -1,1 +1,0 @@
-from numba.cuda.runtime.nrt import rtsys  # noqa: F401
