+++ numba_cuda/numba/cuda/tests/nrt/test_nrt_refct.py
@@ -1,7 +1,7 @@
 import numpy as np
 import unittest
 from numba.tests.support import override_config
-from numba.cuda.runtime import rtsys
+from numba.cuda.memory_management import rtsys
 from numba.cuda.tests.support import EnableNRTStatsMixin
 from numba.cuda.testing import CUDATestCase
